requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
pyautogui>=0.9.53
matplotlib>=3.3.0
scikit-learn>=1.0.0
mss>=6.1.0 